requirements.txt
mysql-connector-python==8.1.0
mcp[cli]==1.6.0
python-dotenv==1.0.0